Extend tests, execute them with github actions and make wendzelnntpd compatible with 64 bit time_t on 32 bit systems

+++ src/db_mysql.c
@@ -200,8 +200,8 @@ db_mysql_post_insert_into_postings(server_cb_inf *inf, char *message_id,
 
 	snprintf(sql_cmd, len - 1,
 		"insert into postings (`msgid`, `date`, `author`, `newsgroups`, `subject`, "
-		"`lines`, `header`) values ('%s', '%li', '%s', '%s', '%s', '%i', '%s');",
-		message_id, ltime, from, ngstrpb, subj, linecount, add_to_hdr);
+		"`lines`, `header`) values ('%s', '%lli', '%s', '%s', '%s', '%i', '%s');",
+		message_id, (long long) ltime, from, ngstrpb, subj, linecount, add_to_hdr);
 
 	if (mysql_real_query(inf->servinf->myhndl, sql_cmd, strlen(sql_cmd)) != 0) {
 		MYSQL_CHK_ERR("mysql_real_query error")
